Test: customer_cannot_destroy_others_orders

+++ app/Http/Controllers/OrderController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Order;
+use Auth;
 use Illuminate\Http\Request;
 
 class OrderController extends Controller
@@ -14,6 +15,10 @@ class OrderController extends Controller
 
     public function destroy(Request $request, Order $order)
     {
-        $order->delete();
+        if (Auth::id() == $order->user_id) {
+            $order->delete();
+        } else {
+            return response()->json(['message' => 'It is not your order!'], 401);
+        }
     }
 }
